Add basic security methods.
Add functions for generating an RSA key, signing a message, and verifying that message. We hash the message with sha1, and sign the hash. Include the signed hash (NOT plaintext hash) and the message. Recipient hashes the message, and decrypts the signed hash, and checks that they are the same.